5 human scouting (#26)
* feat: adds simple scouting for human NPCs

Simple scouting for human NPCs, which allows them to move randomly
in the world looking for the alien intruder.

* feat: adds limit for scouting

Adds a limit to how far from the base an NPC can walk when scouting
for the player.

* refactor: places getter and setter methods at the bottom of the file

* docs: adds documentation to the scouting script

* feat: increases movement of the scouts

Increases movement of the scouts by forcing them to move from a
area if they spend to much time in it. This is done to increase the
area covered by the scouts.

* feat: adds random wait time for scout on destination reached

Adds a timer which allows the scout to wait at a destination for a
given time before receiving a new destination.

* feat: adds stopping mechanic to the scout

* fix: fixes bug with human scout not moving

Fixes the bug where the human scout stops and doesn't start
moving again.