38 expand terrain (#40)
* feat: adds props into the ship

* feat: makes ship static

* feat: adds FOV vision to the AI

* feat: adds player detection and rendered vision

Adds player detection within the vision cone of the AI and
also displays the vision cone which the AI uses to see the
player

* feat: adds hunting mechanic to the AI

* feat: adds simple reputation to the AI

Adds a int which is to represent the reputation that the player
has with the human of a specific faction. This is implemented in
order to make the AI distinguish between a friendly and hostile
player and take action accordingly.

* feat: resize FOV according to obstacles

* fix: solves bug with vision cone

Vision cone is now displayed infront of the AI at all times
and is pointing in the direction which the AI is facing.

* feat: adds the updated enemy as to prefabs

* refactor: removes unused variables and comments

* chore: meta file for npc

* fix: removes bug which results in rendering issues

Removes the bug which made the AI vision field render behind the
character.

* refactor: removes empty folder

* feat: adds crouching which enables player to see AI vision field

* perf: removes debugging gizmo from the vision field

* feat: adds scout spawner

* feat: increases terrain size

Increases the size of the terrain which the player can
walk around on. Also includes terrain modling to the new
terrain.

* feat: adds more walkable terrain and a skybox

* feat: adds play area limitation

Adds invisible walls to limit play area and also adds
unreachable terrain

* fix: removes unused files which cause error message

* feat: adds terrain details with correct shadows

* perf: removes unused assets

* feat: adds ship back into scene

* perf: removes unused assets

* refactor: moves removes prefabs which have been moved

Removes the prefabs which have been moved to a different folder

* chore: removes duplicated folder

* fix: solves issue with missing prefab

---------

Co-authored-by: Jürgen <44424384+JuergenWetzel@users.noreply.github.com>