Inventory (#96)
* Feat movement begin

* Feat player can move

* Feat player rotation

* Fix moving in different directions

* Fixed bugs

* feat: adds laser

* wip: bullet hitbox

* feat: adds missing mesh colliders to base

* fix: fixes bug where player can't access ship inventory

* feat: implements player being able to see AI vision while crouched

* feat: increases quest trigger area

* wip: solving bug with recieving quests

Currently the npc give a quest which resutls in a call to the inventory
scripts function Amount which isn't able to identify "Wood"

* feat: implements mouse into map UI

* wip: adds laser

Adds a laser which the player can trigger by pressing left mouse
button. Needs to create a call to enemy damage when hitting enemy.

* feat: increases detection distance

Increases the detection distance of player to match the laser changes

* feat: adds damage to the laser

* fix: solves bug where AI vision isn't visible when toggled

* Feat Inventory is now correct positioned

* Animations for player (#95)

* Feat movement begin

* Feat player can move

* Feat player rotation

* Fix moving in different directions

* Fixed bugs

* feat: adds laser

* wip: bullet hitbox

* feat: adds missing mesh colliders to base

* fix: fixes bug where player can't access ship inventory

* feat: implements player being able to see AI vision while crouched

* feat: increases quest trigger area

* wip: solving bug with recieving quests

Currently the npc give a quest which resutls in a call to the inventory
scripts function Amount which isn't able to identify "Wood"

* feat: implements mouse into map UI

* wip: adds laser

Adds a laser which the player can trigger by pressing left mouse
button. Needs to create a call to enemy damage when hitting enemy.

* feat: increases detection distance

Increases the detection distance of player to match the laser changes

* feat: adds damage to the laser

* fix: solves bug where AI vision isn't visible when toggled

* fix: Solves bug with quest giver

---------

Co-authored-by: Linus Wallin <linuswallin@live.se>

---------

Co-authored-by: Linus Wallin <linuswallin@live.se>